travel_backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.21.0
asyncpg>=0.27.0
# Use SQLAlchemy >= 1.4 for async support, ideally >= 2.0
SQLAlchemy[asyncio]>=2.0.0
# Pydantic v1 is specified, ensure compatibility or update models for v2 if using >=2.0.0
pydantic>=1.10.0,<2.0.0
python-dotenv>=1.0.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.5 # Needed for OAuth2PasswordRequestForm
# Add email-validator for Pydantic email validation
email-validator>=2.0.0

# Optional for migrations
alembic>=1.10.0